refactor: split license package
