+++ internal/license/manager.go
@@ -12,7 +12,6 @@ import (
 	"net/http"
 	"os"
 	"os/exec"
-	"path/filepath"
 	"runtime"
 	"strings"
 	"sync"
@@ -21,437 +20,6 @@ import (
 	"google.golang.org/api/option"
 	"google.golang.org/api/sheets/v4"
 )
-
-// LogLevel represents different logging levels
-type LogLevel string
-
-const (
-	LogLevelDebug LogLevel = "DEBUG"
-	LogLevelInfo  LogLevel = "INFO"
-	LogLevelWarn  LogLevel = "WARN"
-	LogLevelError LogLevel = "ERROR"
-)
-
-// LogEntry represents a structured log entry
-type LogEntry struct {
-	Timestamp  time.Time `json:"timestamp"`
-	Level      LogLevel  `json:"level"`
-	Action     string    `json:"action"`
-	LicenseKey string    `json:"license_key,omitempty"`
-	MachineID  string    `json:"machine_id,omitempty"`
-	Result     string    `json:"result"`
-	Duration   int64     `json:"duration_ms,omitempty"`
-	Error      string    `json:"error,omitempty"`
-	Details    any       `json:"details,omitempty"`
-	UserAgent  string    `json:"user_agent,omitempty"`
-	IPAddress  string    `json:"ip_address,omitempty"`
-}
-
-// Logger handles structured logging for the license system
-type Logger struct {
-	logFile   *os.File
-	auditFile *os.File
-	mutex     sync.Mutex
-	level     LogLevel
-}
-
-// NewLogger creates a new structured logger
-func NewLogger(logLevel LogLevel) (*Logger, error) {
-	// Create logs directory if it doesn't exist
-	if err := os.MkdirAll("logs", 0755); err != nil {
-		return nil, fmt.Errorf("failed to create logs directory: %v", err)
-	}
-
-	// Open main log file
-	logFile, err := os.OpenFile(
-		filepath.Join("logs", "license.log"),
-		os.O_CREATE|os.O_WRONLY|os.O_APPEND,
-		0644,
-	)
-	if err != nil {
-		return nil, fmt.Errorf("failed to open log file: %v", err)
-	}
-
-	// Open audit log file
-	auditFile, err := os.OpenFile(
-		filepath.Join("logs", "audit.log"),
-		os.O_CREATE|os.O_WRONLY|os.O_APPEND,
-		0644,
-	)
-	if err != nil {
-		logFile.Close()
-		return nil, fmt.Errorf("failed to open audit file: %v", err)
-	}
-
-	return &Logger{
-		logFile:   logFile,
-		auditFile: auditFile,
-		level:     logLevel,
-	}, nil
-}
-
-// Log writes a structured log entry
-func (l *Logger) Log(entry LogEntry) {
-	l.mutex.Lock()
-	defer l.mutex.Unlock()
-
-	// Check if we should log this level
-	if !l.shouldLog(entry.Level) {
-		return
-	}
-
-	// Set timestamp if not provided
-	if entry.Timestamp.IsZero() {
-		entry.Timestamp = time.Now()
-	}
-
-	// Serialize to JSON
-	data, err := json.Marshal(entry)
-	if err != nil {
-		fmt.Printf("Failed to marshal log entry: %v\n", err)
-		return
-	}
-
-	// Write to main log
-	l.logFile.WriteString(string(data) + "\n")
-	l.logFile.Sync()
-
-	// Write security-related actions to audit log
-	if l.isAuditableAction(entry.Action) {
-		l.auditFile.WriteString(string(data) + "\n")
-		l.auditFile.Sync()
-	}
-
-	// Also output to console for development
-	fmt.Printf("[%s] %s: %s\n", entry.Level, entry.Action, entry.Result)
-}
-
-// shouldLog determines if a log level should be written
-func (l *Logger) shouldLog(level LogLevel) bool {
-	levels := map[LogLevel]int{
-		LogLevelDebug: 0,
-		LogLevelInfo:  1,
-		LogLevelWarn:  2,
-		LogLevelError: 3,
-	}
-
-	currentLevel, exists := levels[l.level]
-	if !exists {
-		currentLevel = 1 // Default to INFO
-	}
-
-	entryLevel, exists := levels[level]
-	if !exists {
-		entryLevel = 1
-	}
-
-	return entryLevel >= currentLevel
-}
-
-// isAuditableAction determines if an action should be audited
-func (l *Logger) isAuditableAction(action string) bool {
-	auditableActions := map[string]bool{
-		"license_activation": true,
-		"license_transfer":   true,
-		"license_revocation": true,
-		"license_extension":  true,
-		"validation_failure": true,
-		"security_violation": true,
-		"admin_access":       true,
-	}
-
-	return auditableActions[action]
-}
-
-// Close closes the logger files
-func (l *Logger) Close() error {
-	l.mutex.Lock()
-	defer l.mutex.Unlock()
-
-	var errors []string
-	if err := l.logFile.Close(); err != nil {
-		errors = append(errors, fmt.Sprintf("log file: %v", err))
-	}
-	if err := l.auditFile.Close(); err != nil {
-		errors = append(errors, fmt.Sprintf("audit file: %v", err))
-	}
-
-	if len(errors) > 0 {
-		return fmt.Errorf("failed to close logger: %s", strings.Join(errors, ", "))
-	}
-	return nil
-}
-
-// CacheEntry represents a cached license validation result
-type CacheEntry struct {
-	LicenseInfo LicenseInfo `json:"license_info"`
-	CachedAt    time.Time   `json:"cached_at"`
-	ExpiresAt   time.Time   `json:"expires_at"`
-	HitCount    int         `json:"hit_count"`
-}
-
-// LicenseCache provides intelligent caching for license validations
-type LicenseCache struct {
-	entries   map[string]CacheEntry
-	mutex     sync.RWMutex
-	ttl       time.Duration
-	maxSize   int
-	hitCount  int64
-	missCount int64
-}
-
-// NewLicenseCache creates a new license cache
-func NewLicenseCache(ttl time.Duration, maxSize int) *LicenseCache {
-	cache := &LicenseCache{
-		entries: make(map[string]CacheEntry),
-		ttl:     ttl,
-		maxSize: maxSize,
-	}
-
-	// Start cleanup goroutine
-	go cache.cleanup()
-
-	return cache
-}
-
-// Get retrieves a license from cache
-func (c *LicenseCache) Get(licenseKey string) (*LicenseInfo, bool) {
-	c.mutex.RLock()
-	defer c.mutex.RUnlock()
-
-	entry, exists := c.entries[licenseKey]
-	if !exists || time.Now().After(entry.ExpiresAt) {
-		c.missCount++
-		return nil, false
-	}
-
-	// Update hit count
-	entry.HitCount++
-	c.entries[licenseKey] = entry
-	c.hitCount++
-
-	return &entry.LicenseInfo, true
-}
-
-// Set stores a license in cache
-func (c *LicenseCache) Set(licenseKey string, info LicenseInfo) {
-	c.mutex.Lock()
-	defer c.mutex.Unlock()
-
-	// Evict old entries if cache is full
-	if len(c.entries) >= c.maxSize {
-		c.evictOldest()
-	}
-
-	c.entries[licenseKey] = CacheEntry{
-		LicenseInfo: info,
-		CachedAt:    time.Now(),
-		ExpiresAt:   time.Now().Add(c.ttl),
-		HitCount:    0,
-	}
-}
-
-// Invalidate removes a license from cache
-func (c *LicenseCache) Invalidate(licenseKey string) {
-	c.mutex.Lock()
-	defer c.mutex.Unlock()
-	delete(c.entries, licenseKey)
-}
-
-// GetStats returns cache statistics
-func (c *LicenseCache) GetStats() map[string]interface{} {
-	c.mutex.RLock()
-	defer c.mutex.RUnlock()
-
-	totalRequests := c.hitCount + c.missCount
-	hitRatio := float64(0)
-	if totalRequests > 0 {
-		hitRatio = float64(c.hitCount) / float64(totalRequests)
-	}
-
-	return map[string]interface{}{
-		"entries":     len(c.entries),
-		"max_size":    c.maxSize,
-		"hit_count":   c.hitCount,
-		"miss_count":  c.missCount,
-		"hit_ratio":   hitRatio,
-		"ttl_seconds": c.ttl.Seconds(),
-	}
-}
-
-// evictOldest removes the oldest cache entry
-func (c *LicenseCache) evictOldest() {
-	var oldestKey string
-	var oldestTime time.Time
-
-	for key, entry := range c.entries {
-		if oldestKey == "" || entry.CachedAt.Before(oldestTime) {
-			oldestKey = key
-			oldestTime = entry.CachedAt
-		}
-	}
-
-	if oldestKey != "" {
-		delete(c.entries, oldestKey)
-	}
-}
-
-// cleanup periodically removes expired entries
-func (c *LicenseCache) cleanup() {
-	ticker := time.NewTicker(5 * time.Minute)
-	defer ticker.Stop()
-
-	for range ticker.C {
-		c.mutex.Lock()
-		now := time.Now()
-		for key, entry := range c.entries {
-			if now.After(entry.ExpiresAt) {
-				delete(c.entries, key)
-			}
-		}
-		c.mutex.Unlock()
-	}
-}
-
-// SecurityManager handles rate limiting and security protection
-type SecurityManager struct {
-	attemptCounts   map[string]int
-	lastAttempts    map[string]time.Time
-	blockedIPs      map[string]time.Time
-	mutex           sync.RWMutex
-	maxAttempts     int
-	blockDuration   time.Duration
-	windowDuration  time.Duration
-	cleanupInterval time.Duration
-	logger          *Logger
-}
-
-// NewSecurityManager creates a new security manager
-func NewSecurityManager(maxAttempts int, blockDuration, windowDuration time.Duration, logger *Logger) *SecurityManager {
-	sm := &SecurityManager{
-		attemptCounts:   make(map[string]int),
-		lastAttempts:    make(map[string]time.Time),
-		blockedIPs:      make(map[string]time.Time),
-		maxAttempts:     maxAttempts,
-		blockDuration:   blockDuration,
-		windowDuration:  windowDuration,
-		cleanupInterval: 5 * time.Minute,
-		logger:          logger,
-	}
-
-	// Start cleanup goroutine
-	go sm.cleanup()
-
-	return sm
-}
-
-// IsBlocked checks if an identifier is currently blocked
-func (s *SecurityManager) IsBlocked(identifier string) bool {
-	s.mutex.RLock()
-	defer s.mutex.RUnlock()
-
-	if blockTime, exists := s.blockedIPs[identifier]; exists {
-		if time.Since(blockTime) < s.blockDuration {
-			return true
-		}
-		delete(s.blockedIPs, identifier)
-	}
-	return false
-}
-
-// RecordAttempt records a license operation attempt
-func (s *SecurityManager) RecordAttempt(identifier string, success bool) bool {
-	s.mutex.Lock()
-	defer s.mutex.Unlock()
-
-	now := time.Now()
-
-	if success {
-		// Clear attempts on success
-		delete(s.attemptCounts, identifier)
-		delete(s.lastAttempts, identifier)
-		return true
-	}
-
-	// Check if this is within the time window
-	if lastAttempt, exists := s.lastAttempts[identifier]; exists {
-		if now.Sub(lastAttempt) > s.windowDuration {
-			// Reset counter if outside window
-			s.attemptCounts[identifier] = 1
-		} else {
-			s.attemptCounts[identifier]++
-		}
-	} else {
-		s.attemptCounts[identifier] = 1
-	}
-
-	s.lastAttempts[identifier] = now
-
-	// Check if we should block
-	if s.attemptCounts[identifier] >= s.maxAttempts {
-		s.blockedIPs[identifier] = now
-
-		// Log security violation
-		if s.logger != nil {
-			s.logger.Log(LogEntry{
-				Level:     LogLevelWarn,
-				Action:    "security_violation",
-				Result:    "IP blocked due to too many failed attempts",
-				IPAddress: identifier,
-				Details: map[string]interface{}{
-					"attempt_count": s.attemptCounts[identifier],
-					"max_attempts":  s.maxAttempts,
-				},
-			})
-		}
-
-		return false
-	}
-
-	return true
-}
-
-// GetStats returns security statistics
-func (s *SecurityManager) GetStats() map[string]interface{} {
-	s.mutex.RLock()
-	defer s.mutex.RUnlock()
-
-	return map[string]interface{}{
-		"active_attempts": len(s.attemptCounts),
-		"blocked_ips":     len(s.blockedIPs),
-		"max_attempts":    s.maxAttempts,
-		"block_duration":  s.blockDuration.String(),
-		"window_duration": s.windowDuration.String(),
-	}
-}
-
-// cleanup periodically removes old entries
-func (s *SecurityManager) cleanup() {
-	ticker := time.NewTicker(s.cleanupInterval)
-	defer ticker.Stop()
-
-	for range ticker.C {
-		s.mutex.Lock()
-		now := time.Now()
-
-		// Clean up old attempts
-		for identifier, lastAttempt := range s.lastAttempts {
-			if now.Sub(lastAttempt) > s.windowDuration {
-				delete(s.attemptCounts, identifier)
-				delete(s.lastAttempts, identifier)
-			}
-		}
-
-		// Clean up expired blocks
-		for identifier, blockTime := range s.blockedIPs {
-			if now.Sub(blockTime) > s.blockDuration {
-				delete(s.blockedIPs, identifier)
-			}
-		}
-
-		s.mutex.Unlock()
-	}
-}
 
 // LicenseInfo represents license data
 type LicenseInfo struct {
